requirements.txt
llama-index
google-generativeai
llama-index-llms-gemini
pypdf
python-dotenv
IPython
llama-index-embeddings-gemini
streamlit
html2text
llama-index-readers-web
ipykernel
nbconvert
docx2txt
#-e . //to download up the setup.py file content as well